requirements.txt
streamlit>=1.28.0
gspread>=5.10.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=1.5.0
altair>=5.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
